refactor part 2

+++ NuklearC.Native/Types/nk_plugin_alloc_t.cs
@@ -4,5 +4,5 @@ using System.Runtime.InteropServices;
 namespace NuklearC.Native.Types
 {
     [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
-    public delegate IntPtr nk_plugin_alloc_t(nk_handle handle, IntPtr old, IntPtr nk_size);
+    public delegate IntPtr nk_plugin_alloc_t(nk_handle handle, IntPtr old, int nk_size);
 }
